% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FarringtonManning.R
\docType{methods}
\name{adjusted_alpha,FarringtonManning-method}
\alias{adjusted_alpha,FarringtonManning-method}
\title{Calculation of the Adjusted Alpha for the Farrington-Manning}
\usage{
\S4method{adjusted_alpha}{FarringtonManning}(design, n1, nuisance,
  precision = 0.001, recalculation, allocation = c("exact",
  "approximate"), ...)
}
\arguments{
\item{n1}{Either the total sample size or sample size of the first stage}

\item{nuisance}{A vector of nuisance parameters}
}
\description{
Calculates the adjusted alpha that is necessary to maintain the nominimal type 1
error rate for the fixed sample design and the internal pilot study design in the
Farrington-Manning test.
}
