% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R, R/Student.R, R/ChiSquare.R,
%   R/FarringtonManning.R
\name{n_dist}
\alias{n_dist}
\alias{n_dist,Student-method}
\alias{n_dist,ChiSquare-method}
\alias{n_dist,FarringtonManning-method}
\title{Distribution of the Sample Size}
\usage{
n_dist(design, n1, nuisance, summary = TRUE, plot = FALSE, ...)

\S4method{n_dist}{Student}(
  design,
  n1,
  nuisance,
  summary = TRUE,
  plot = FALSE,
  iters = 10000,
  seed = NULL,
  range = 0,
  allocation = c("approximate", "exact"),
  ...
)

\S4method{n_dist}{ChiSquare}(
  design,
  n1,
  nuisance,
  summary,
  plot,
  allocation = c("exact", "approximate"),
  ...
)

\S4method{n_dist}{FarringtonManning}(
  design,
  n1,
  nuisance,
  summary,
  plot,
  allocation = c("exact", "approximate"),
  ...
)
}
\arguments{
\item{design}{object of class \code{TestStatistc} created by \code{setup}}

\item{n1}{number of patients that are recruited before the sample size
is recalculated}

\item{nuisance}{nuisance parameter that is estimated at the interim analysis}

\item{summary}{logical - is a summary of the sample size distribution desired?
Otherwise, a vector with sample sizes is returned.}

\item{plot}{logical - should a plot be created?}

\item{...}{further optional arguments}

\item{iters}{number of simulation iterations}

\item{seed}{random seed for simulation}

\item{range}{this determines how far the plot whiskers extend out from the box.
If range is positive, the whiskers extend to the most extreme data point
which is no more than range times the interquartile range from the box.
A value of zero causes the whiskers to extend to the data extremes.}

\item{allocation}{should the allocation ratio be hold exactly (option 'exact') or approximately (option 'approximate')?}
}
\description{
Distribution of the Sample Size
}
\examples{
d <- setupStudent(alpha = .025, beta = .2, r = 1, delta = 3.5, delta_NI = 0,
                  alternative = "greater", n_max = 156)
n_dist(d, n1 = 20, nuisance = 5.5, summary = TRUE, plot = FALSE, seed = 2020)

}
