% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChiSquare.R
\name{n_fix,ChiSquare-method}
\alias{n_fix,ChiSquare-method}
\title{Sample Size Calculation for the Chi-Squared Test}
\usage{
\S4method{n_fix}{ChiSquare}(
  design,
  nuisance,
  variance = c("heterogeneous", "homogeneous"),
  rounded = TRUE,
  ...
)
}
\arguments{
\item{design}{an object of class \code{ChiSquare} created by \code{setupChiSquare()}.}

\item{nuisance}{the overall response rate.}

\item{variance}{a character string indicating whether the "\code{heterogenous}" (default)
or the "\code{homogeneous}" variance formula should be used.}
}
\value{

}
\description{
Calculates the sample size for the chi-squared test for the corresponding
one-stage design without sample size recalculation.
}
