% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FarringtonManning.R
\name{n_fix,FarringtonManning-method}
\alias{n_fix,FarringtonManning-method}
\title{Sample Size Calculation for the Farrington-Manning Test}
\usage{
\S4method{n_fix}{FarringtonManning}(design, nuisance, rounded = TRUE, ...)
}
\arguments{
\item{design}{an object of class \code{FarringtonManning} created
by \code{setupFarringtonManning()}.}

\item{nuisance}{the overall response rate.}

\item{rounded}{}
}
\description{
Calculates the sample size for the Farrington-Manning test for the corresponding
one-stage design without sample size recalculation.
}
