% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R, R/Student.R
\name{n_fix}
\alias{n_fix}
\alias{n_fix,Student-method}
\title{Fixed Sample Size}
\usage{
n_fix(design, nuisance, ...)

\S4method{n_fix}{Student}(design, nuisance, ...)
}
\arguments{
\item{design}{test statistic object}

\item{nuisance}{nuisance parameter}

\item{...}{further optional arguments}
}
\description{
Returns the sample size for the corresponding one-stage design without
sample size recalculation.
}
\examples{
d <- setupStudent(alpha = .025, beta = .2, r = 1, delta = 3.5, delta_NI = 0,
                  alternative = "greater", n_max = 156)
n_fix(design = d, nuisance = 5.5)

}
