% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChiSquare.R
\name{pow,ChiSquare-method}
\alias{pow,ChiSquare-method}
\title{Calculation of the Power of the Chi-Squared Test}
\usage{
\S4method{pow}{ChiSquare}(
  design,
  n1,
  nuisance,
  recalculation,
  allocation = c("exact", "approximate"),
  ...
)
}
\arguments{
\item{n1}{Either the total sample size (if \code{design} is \code{"fixed"}) or
sample size of the first stage (if \code{design} is \code{"ips"})}

\item{nuisance}{the overall response rate.}

\item{recalculation}{logical: should the sample size be recalcuated after n1
patients are recruited?}

\item{allocation}{should the allocation ratio be hold exactly (option 'exact') or approximately (option 'approximate')?}

\item{...}{further optional arguments}
}
\value{

}
\description{
Calculation of the power of the chi-squared test for the fixed sample design and the
internal pilot study design.
}
