% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{sample_size_dist}
\alias{sample_size_dist}
\title{Distribution of the Sample Size}
\usage{
sample_size_dist(design, n1, nuisance, recalculation, ...)
}
\arguments{
\item{n1}{number of patients that are recruited before the sample size
is recalculated}

\item{nuisance}{nuisance parameter that is estimated at the interim analysis}

\item{recalculation}{logical: should the sample size be recalcuated after n1
patients are recruited?}

\item{...}{further optional arguments}

\item{s}{object of class \code{TestStatistc} created by \code{setup}}
}
\description{
Distribution of the Sample Size
}
