% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_statistics.R
\name{setupChiSquare}
\alias{setupChiSquare}
\title{Setup a Chi^2 test}
\usage{
setupChiSquare(alpha, beta, r = 1, delta, n_max = Inf, ...)
}
\arguments{
\item{alpha}{one-sided type I error rate}

\item{beta}{type II error rate}

\item{r}{allocation ratio between experimental and control group}

\item{delta}{difference of effect size between alternative and null hypothesis}

\item{n_max}{maximal overall sample size}

\item{...}{further optional arguments}
}
\description{
This function creates an object of class \code{ChiSquare} that can
be used for sample size recalculation.
}
\details{
For non-inferiority trials use the function \code{\link{setupFarringtionManning}}.
}
