% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FarringtonManning.R
\name{toer,FarringtonManning-method}
\alias{toer,FarringtonManning-method}
\title{Calculation of the Actual Level of the Chi-Squared Test}
\usage{
\S4method{toer}{FarringtonManning}(
  design,
  n1,
  nuisance,
  recalculation,
  allocation = c("exact", "approximate"),
  ...
)
}
\arguments{
\item{design}{an object of class \code{FarringtonManning} created
by \code{setupFarringtonManning()}.}

\item{n1}{Either the total sample size (if \code{design} is \code{"fixed"}) or
sample size of the first stage (if \code{design} is \code{"ips"})}

\item{nuisance}{the overall response rate.}

\item{recalculation}{logical: should the sample size be recalcuated after n1
patients are recruited?}

\item{allocation}{should the allocation ratio be hold exactly (option 'exact') or approximately (option 'approximate')?}

\item{...}{further optional arguments}
}
\description{
Calculation of the actual level of the chi-squared test for the fixed sample design and the
internal pilot study design.
}
