% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R, R/Student.R
\name{adjusted_alpha}
\alias{adjusted_alpha}
\alias{adjusted_alpha,Student-method}
\title{Adjusted level of significance}
\usage{
adjusted_alpha(design, n1, nuisance, ...)

\S4method{adjusted_alpha}{Student}(design, n1, nuisance, tol, iters = 10000, seed = NULL, ...)
}
\arguments{
\item{design}{object of class \code{TestStatistc} created by \code{setup}}

\item{n1}{number of patients that are recruited before the sample size
is recalculated}

\item{nuisance}{nuisance parameter that is estimated at the interim analysis}

\item{...}{further optional arguments}

\item{tol}{desired absolute tolerance}

\item{iters}{number of simulation iterations}

\item{seed}{random seed for simulation}
}
\description{
This method returns an adjusted significance level that can be used
such that the actual type I error rate is protected.
}
\details{
In the case of the Student's t-test, the adjusted alpha is calculated
using the algorithm by Kieser and Friede (2000):
"Re-calculating the sample size in internal pilot study designs
with control of the type I error rate"
}
