% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_statistics.R, R/FarringtonManning.R
\docType{class}
\name{FarringtonManning-class}
\alias{FarringtonManning-class}
\alias{FarringtonManning}
\alias{setupFarringtonManning}
\alias{n_fix,FarringtonManning-method}
\alias{toer,FarringtonManning-method}
\alias{pow,FarringtonManning-method}
\alias{adjusted_alpha,FarringtonManning-method}
\alias{n_dist,FarringtonManning-method}
\title{Farrington Manning test}
\usage{
setupFarringtonManning(alpha, beta, r = 1, delta, delta_NI, n_max = Inf, ...)

\S4method{n_fix}{FarringtonManning}(design, nuisance, rounded = TRUE, ...)

\S4method{toer}{FarringtonManning}(
  design,
  n1,
  nuisance,
  recalculation,
  allocation = c("exact", "approximate"),
  ...
)

\S4method{pow}{FarringtonManning}(
  design,
  n1,
  nuisance,
  recalculation,
  allocation = c("exact", "approximate"),
  ...
)

\S4method{adjusted_alpha}{FarringtonManning}(
  design,
  n1,
  nuisance,
  nuis_ass,
  precision = 0.001,
  gamma = 0,
  recalculation,
  allocation = c("exact", "approximate"),
  ...
)

\S4method{n_dist}{FarringtonManning}(
  design,
  n1,
  nuisance,
  summary,
  plot,
  allocation = c("exact", "approximate"),
  ...
)
}
\arguments{
\item{alpha}{one-sided type I error rate}

\item{beta}{type II error rate}

\item{r}{allocation ratio between experimental and control group}

\item{delta}{difference of effect size between alternative and null hypothesis}

\item{delta_NI}{non-inferiority margin}

\item{n_max}{maximal overall sample size}

\item{...}{further optional arguments}

\item{design}{Object of class \code{FarringtonManning} created
by \code{setupFarringtonManning}.}

\item{nuisance}{Value of the nuisance parameter. For the
Farrington-Manning test this is the overall response rate.}

\item{rounded}{Whether the calculated sample size should be rounded up such that
the allocation ratio is preserved.}

\item{n1}{Either the sample size of the first stage (if
\code{recalculation = TRUE} or the toal sample size (if
\code{recalculation = FALSE}).}

\item{recalculation}{Logical: Should the sample size be recalcuated after n1
patients are recruited?}

\item{allocation}{Whether the allocation ratio should be preserved
exactly (\code{exact}) or approximately (\code{approximate}).}

\item{nuis_ass}{If \code{recalculation = FALSE} this is the value for
the overall response rate that is used to calculate the sample size
for the adjusted significance level.}

\item{precision}{Value by which the nominal type 1 error rate is
reduced in each iteration until the nominal type 1 error rate is
preserved.}

\item{gamma}{If \code{gamma > 0} then the significance level is adjusted
such that the actual level is at most \code{alpha - gamma}. This is
necessary to maintain the nomininal significance level if a confidence
interval approach proposed by Friede & Kieser (2011) is used.}

\item{summary}{Logical. If \code{TRUE} (default) a summary of the sample
size distribution is printed. If \code{FALSE} all sample sizes are
printed.}

\item{plot}{Logical. Should a plot of the sample size distribution
be provided?}
}
\description{
This class implements a Farrington-Manning test for non-inferiority
trials. A trial with binary outcomes in two groups \code{T} and
\code{C} is assumed.

The function \code{\link{setupFarringtonManning}} creates an object of
\code{\link{FarringtonManning}}.
}
\examples{
design <- setupFarringtonManning(alpha = .025, beta = .2, r = 1, delta = 0,
delta_NI = .15)
}
