% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_statistics.R
\name{setupFarringtonManning}
\alias{setupFarringtonManning}
\title{Setup a Farrington Manning test}
\usage{
setupFarringtonManning(alpha, beta, r = 1, delta, delta_NI,
  n_max = Inf, ...)
}
\arguments{
\item{alpha}{one-sided type I error rate}

\item{beta}{type II error rate}

\item{r}{allocation ratio between experimental and control group}

\item{delta}{difference of effect size between alternative and null hypothesis}

\item{delta_NI}{non-inferiority margin}

\item{n_max}{maximal overall sample size}

\item{...}{further optional arguments}
}
\description{
This function creates an object of class \code{\link{FarringtonManning-class}}
that can be used for sample size recalculation in non-inferiority trials with
binary endpoints.
}
