% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R, R/Student.R, R/ChiSquare.R,
%   R/FarringtonManning.R
\name{n_dist}
\alias{n_dist}
\alias{n_dist,Student-method}
\alias{n_dist,ChiSquare-method}
\alias{n_dist,FarringtonManning-method}
\title{Distribution of the Sample Size}
\usage{
n_dist(design, n1, nuisance, summary = TRUE, plot = FALSE, ...)

\S4method{n_dist}{Student}(
  design,
  n1,
  nuisance,
  summary = TRUE,
  plot = FALSE,
  iters = 10000,
  seed = NULL,
  range = 0,
  allocation = c("approximate", "exact"),
  ...
)

\S4method{n_dist}{ChiSquare}(
  design,
  n1,
  nuisance,
  summary,
  plot,
  allocation = c("exact", "approximate"),
  ...
)

\S4method{n_dist}{FarringtonManning}(
  design,
  n1,
  nuisance,
  summary,
  plot,
  allocation = c("exact", "approximate"),
  ...
)
}
\arguments{
\item{design}{object of class \code{TestStatistc} created by \code{setup}}

\item{n1}{number of patients that are recruited before the sample size
is recalculated}

\item{nuisance}{nuisance parameter that is estimated at the interim analysis}

\item{summary}{logical - is a summary of the sample size distribution desired?
Otherwise, a vector with sample sizes is returned.}

\item{plot}{logical - should a plot be created?}

\item{...}{further optional arguments}

\item{iters}{number of simulation iterations}

\item{seed}{random seed for simulation}
}
\description{
Distribution of the Sample Size
}
