% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R, R/Student.R
\name{n_fix}
\alias{n_fix}
\alias{n_fix,Student-method}
\title{Fixed Sample Size}
\usage{
n_fix(design, nuisance, ...)

\S4method{n_fix}{Student}(design, nuisance, ...)
}
\arguments{
\item{design}{test statistic object}

\item{nuisance}{nuisance parameter}

\item{...}{further optional arguments}
}
\description{
Returns the sample size for the corresponding one-stage design without
sample size recalculation.
}
