% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{n_fix}
\alias{n_fix}
\title{Fixed sample size}
\usage{
n_fix(s, nuisance, ...)
}
\arguments{
\item{s}{test statistic object}

\item{nuisance}{nuisance parameter}
}
\description{
Returns the sample size for the corresponding one-stage design without
sample size recalculation.
}
