% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FarringtonManning.R
\docType{methods}
\name{pow,FarringtonManning-method}
\alias{pow,FarringtonManning-method}
\title{Calculation of the Power of the Farrington-Manning test}
\usage{
\S4method{pow}{FarringtonManning}(design, n1, nuisance, recalculation,
  allocation = c("exact", "approximate"), ...)
}
\arguments{
\item{n1}{Either the total sample size (if \code{design} is \code{"fixed"}) or
sample size of the first stage (if \code{design} is \code{"ips"})}

\item{nuisance}{the overall response rate.}

\item{allocation}{}
}
\description{
Calculation of the power of the Farrington-Manning test for the fixed sample design
and the internal pilot study design.
}
