% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R, R/Student.R
\name{pow}
\alias{pow}
\alias{pow,Student-method}
\title{Power}
\usage{
pow(s, n1, nuisance, ...)

\S4method{pow}{Student}(s, n1, nuisance, recalculation = TRUE, iters, ...)
}
\arguments{
\item{n1}{number of patients that are recruited before the sample size
is recalculated}

\item{nuisance}{nuisance parameter that is estimated at the interim analysis}

\item{design}{trial design created by \code{\link{setupStudent}}}

\item{recalculate}{logical: should the sample size be recalcuated after n1
patients are recruited?}
}
\description{
Power
}
