% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R, R/Student.R
\docType{methods}
\name{pow}
\alias{pow}
\alias{pow,Student-method}
\title{Power}
\usage{
pow(design, n1, nuisance, recalculation, ...)

\S4method{pow}{Student}(design, n1, nuisance, recalculation = TRUE,
  iters, ...)
}
\arguments{
\item{n1}{number of patients that are recruited before the sample size
is recalculated}

\item{nuisance}{nuisance parameter that is estimated at the interim analysis}

\item{recalculation}{logical: should the sample size be recalcuated after n1
patients are recruited?}

\item{...}{further optional arguments}

\item{iters}{number of simulation iterations}

\item{s}{object of class \code{TestStatistc} created by \code{setup}}
}
\description{
Power
}
