% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{sample_size_dist}
\alias{sample_size_dist}
\title{Distribution of the sample size}
\usage{
sample_size_dist(s, n1, nuisance, recalculation, ...)
}
\arguments{
\item{s}{object of class \code{TestStatistc} created by \code{setup}}

\item{n1}{number of patients that are recruited before the sample size
is recalculated}

\item{nuisance}{nuisance parameter that is estimated at the interim analysis}

\item{recalculation}{logical: should the sample size be recalcuated after n1
patients are recruited?}

\item{...}{further optional arguments}
}
\description{
Distribution of the sample size
}
