% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Student.R
\name{simulation}
\alias{simulation}
\title{Simulate rejection probability and sample size for Student's t-test}
\usage{
simulation(
  s,
  n1,
  nuisance,
  recalculation = TRUE,
  Delta_star,
  iters = 1000,
  ...
)
}
\arguments{
\item{s}{object of class \code{TestStatistc} created by \code{setup}}

\item{n1}{number of patients that are recruited before the sample size
is recalculated}

\item{nuisance}{nuisance parameter that is estimated at the interim analysis}

\item{recalculation}{logical: should the sample size be recalcuated after n1
patients are recruited?}

\item{Delta_star}{effect measure under which the rejection probabilities are computed}

\item{iters}{number of simulation iterations}

\item{...}{further optional arguments}
}
\description{
This function simulates the probability that a test defined by
\code{\link{setupStudent}} rejects the null hypothesis.
Note that here the nuisance parameter \code{nuisance} is the variance
of the outcome variable sigma^2.
}
\details{
The implementation follows the algorithm in Lu (2019):
Distribution of the two-sample t-test statistic following blinded
sample size re-estimation.
Pharmaceutical Statistics 15: 208-215.
}
