% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_statistics.R, R/Student.R
\docType{class}
\name{Student-class}
\alias{Student-class}
\alias{Student}
\alias{setupStudent}
\alias{toer,Student-method}
\alias{pow,Student-method}
\alias{n_dist,Student-method}
\alias{n_fix,Student-method}
\alias{adjusted_alpha,Student-method}
\title{Student's t test}
\usage{
setupStudent(
  alpha,
  beta,
  r = 1,
  delta,
  delta_NI = 0,
  alternative = c("greater", "smaller"),
  n_max = Inf,
  ...
)

\S4method{toer}{Student}(
  design,
  n1,
  nuisance,
  recalculation = TRUE,
  iters = 10000,
  seed = NULL,
  allocation = c("approximate", "exact"),
  ...
)

\S4method{pow}{Student}(
  design,
  n1,
  nuisance,
  recalculation = TRUE,
  iters = 10000,
  seed = NULL,
  allocation = c("approximate", "exact"),
  ...
)

\S4method{n_dist}{Student}(
  design,
  n1,
  nuisance,
  summary = TRUE,
  plot = FALSE,
  iters = 10000,
  seed = NULL,
  range = 0,
  allocation = c("approximate", "exact"),
  ...
)

\S4method{n_fix}{Student}(design, nuisance, ...)

\S4method{adjusted_alpha}{Student}(design, n1, nuisance, tol, iters = 10000, seed = NULL, ...)
}
\arguments{
\item{alpha}{one-sided type I error rate}

\item{beta}{type II error rate}

\item{r}{allocation ratio between experimental and control group}

\item{delta}{difference of effect size between alternative and null hypothesis}

\item{delta_NI}{non-inferiority margin}

\item{alternative}{character value indicating if the alternative contains greater
(option "greater") or smaller (option "smaller") values than the null hypothesis}

\item{n_max}{maximal overall sample size}

\item{...}{further optional arguments}

\item{iters}{number of simulation iterations}

\item{seed}{random seed for simulation}

\item{allocation}{should the allocation ratio be hold exactly (option 'exact') or approximately (option 'approximate')?}

\item{range}{this determines how far the plot whiskers extend out from the box.
If range is positive, the whiskers extend to the most extreme data point
which is no more than range times the interquartile range from the box.
A value of zero causes the whiskers to extend to the data extremes.}

\item{tol}{desired absolute tolerance}
}
\description{
This class implements Student's t-test for superiority and non-inferiority
tests.
A trial with continuous outcomes of the two groups \code{T} and \code{C}
is assumed.
If \code{alternative == "greater"} the null hypothesis for the
mean difference
\ifelse{html}{\out{&Delta; = &mu;<sub>T</sub> - &mu;<sub>C</sub>}}{\eqn{\Delta = \mu_T - \mu_C}}
is
\ifelse{html}{\out{<p>H<sub>0</sub>: &Delta; &le; -&delta;<sub>NI</sub>  vs.  H<sub>1</sub>: &Delta; > -&delta;<sub>NI</sub>.</p>}}{\deqn{H_0: \Delta \leq -\delta_{NI}  vs.  H_1: \Delta > -\delta_{NI}.}}
Here, \ifelse{html}{\out{&delta;<sub>NI</sub> >0}}{\eqn{\delta_{NI} > 0}} denotes the non-inferiority margin.
If \code{alternative=="smaller"}, the direction of the effect is changed.

The function \code{setupStudent} creates an object of class
\code{\link{Student}} that can be used for sample size recalculation.
}
\details{
The notation is based on the paper of Lu (2019):
Distribution of the two-sample t-test statistic following blinded
sample size re-estimation. Pharmaceutical Statistics 15: 208-215.

In the case of the Student's t-test, the adjusted alpha is calculated
using the algorithm by Kieser and Friede (2000):
"Re-calculating the sample size in internal pilot study designs
with control of the type I error rate"
}
\examples{
d <- setupStudent(alpha = .025, beta = .2, r = 1, delta = 3.5, delta_NI = 0,
                   alternative = "greater", n_max = 156)
}
