% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_statistics.R
\docType{class}
\name{TestStatistic-class}
\alias{TestStatistic-class}
\title{Test statistics}
\description{
In \code{blindrecalc}, different test statistics are implemented.
Currently, those are Student's t-test for superiority and non-inferiorty
for continuous outcomes, the Chi^2-test for superiority tests for binary
outcomes and the Farrington Manning test for non-inferiority tests for
binary outcomes.
}
