% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChiSquare.R
\docType{methods}
\name{toer,ChiSquare-method}
\alias{toer,ChiSquare-method}
\title{Calculation of the Actual Level of the Chi-Squared Test}
\usage{
\S4method{toer}{ChiSquare}(design, n1, nuisance, recalculation,
  allocation = c("exact", "approximate"), ...)
}
\arguments{
\item{n1}{Either the total sample size (if \code{design} is \code{"fixed"}) or
sample size of the first stage (if \code{design} is \code{"ips"})}

\item{nuisance}{the overall response rate.}

\item{allocation}{}
}
\description{
Calculation of the actual level of the chi-squared test for the fixed sample design and the
internal pilot study design.
}
