% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R, R/Student.R
\name{toer}
\alias{toer}
\alias{toer,Student-method}
\title{Type I error rate}
\usage{
toer(s, n1, nuisance, ...)

\S4method{toer}{Student}(s, n1, nuisance, recalculation = TRUE, iters, ...)
}
\arguments{
\item{n1}{number of patients that are recruited before the sample size
is recalculated}

\item{nuisance}{nuisance parameter that is estimated at the interim analysis}

\item{design}{trial design created by \code{\link{setupStudent}}}

\item{recalculate}{logical: should the sample size be recalcuated after n1
patients are recruited?}
}
\value{
one type I error rate value for every nuisance parameter
}
\description{
Computes the type I error rate of a design with blinded sample size recalculation
by simulation (in the continuous case) or exactly (in the binary case)
for one or several values of the nuisance parameter.
}
