% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R, R/Student.R
\name{toer}
\alias{toer}
\alias{toer,Student-method}
\title{Type I Error Rate}
\usage{
toer(design, n1, nuisance, recalculation, ...)

\S4method{toer}{Student}(
  design,
  n1,
  nuisance,
  recalculation = TRUE,
  iters = 10000,
  seed = NULL,
  allocation = c("approximate", "exact"),
  ...
)
}
\arguments{
\item{design}{object of class \code{TestStatistc} created by \code{setup}}

\item{n1}{number of patients that are recruited before the sample size
is recalculated}

\item{nuisance}{nuisance parameter that is estimated at the interim analysis}

\item{recalculation}{logical: should the sample size be recalcuated after n1
patients are recruited?}

\item{...}{further optional arguments}

\item{iters}{number of simulation iterations}

\item{seed}{random seed for simulation}

\item{allocation}{should the allocation ratio be hold exactly (option 'exact') or approximately (option 'approximate')?}
}
\value{
one type I error rate value for every nuisance parameter
}
\description{
Computes the type I error rate of a design with blinded sample size recalculation
by simulation (in the continuous case) or exactly (in the binary case)
for one or several values of the nuisance parameter.
}
